+++ packages/reference/src/components/MissingEntityCard/MissingEntityCard.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { Card, SectionHeading, IconButton } from '@contentful/f36-components';
+
+import { Card, SectionHeading, IconButton, Flex } from '@contentful/f36-components';
 import { CloseIcon } from '@contentful/f36-icons';
 
 import { ContentEntityType } from '../../types';
@@ -13,23 +14,23 @@ export function MissingEntityCard(props: {
 }) {
   return (
     <Card className={styles.card}>
-      <div className={props.asSquare ? styles.squareCard : ''}>
-        <SectionHeading marginBottom="none">
-          {props.entityType === 'Entry' && 'Entry is missing or inaccessible'}
-          {props.entityType === 'Asset' && 'Asset is missing or inaccessible'}
-        </SectionHeading>
-      </div>
-      {!props.isDisabled && props.onRemove && (
-        <IconButton
-          variant="transparent"
-          icon={<CloseIcon variant="muted" />}
-          className={styles.close}
-          aria-label="Delete"
-          onClick={() => {
-            props.onRemove && props.onRemove();
-          }}
-        />
-      )}
+      <Flex alignItems="center" justifyContent="space-between">
+        <div className={props.asSquare ? styles.squareCard : ''}>
+          <SectionHeading marginBottom="none">
+            {props.entityType} is missing or inaccessible
+          </SectionHeading>
+        </div>
+        {!props.isDisabled && props.onRemove && (
+          <IconButton
+            variant="transparent"
+            icon={<CloseIcon variant="muted" />}
+            aria-label="Delete"
+            onClick={() => {
+              props.onRemove && props.onRemove();
+            }}
+          />
+        )}
+      </Flex>
     </Card>
   );
 }