+++ packages/reference/src/components/MissingEntityCard/styles.ts
@@ -1,5 +1,4 @@
 import { css } from 'emotion';
-import tokens from '@contentful/f36-tokens';
 
 export const card = css({
   position: 'relative',
@@ -11,10 +10,4 @@ export const squareCard = css({
   width: '135px',
   height: '160px',
   textAlign: 'center',
-});
-
-export const close = css({
-  position: 'absolute',
-  top: tokens.spacingS,
-  right: tokens.spacingS,
 });